requirements.txt
# ─────────────────────────────────────────────────────────────────
# INF8225 — MK-UNet  (Équipe 5)
#
# IMPORTANT — PyTorch AMD (ROCm) à installer AVANT ce fichier :
#   pip install torch torchvision torchaudio \
#       --index-url https://download.pytorch.org/whl/rocm6.2
#   (remplace rocm6.2 par ta version : rocm-smi --version)
#
# Ensuite : pip install -r requirements.txt
# ─────────────────────────────────────────────────────────────────

timm>=0.9.0
albumentations>=1.3.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
tqdm>=4.65.0
scipy>=1.11.0
pandas>=2.0.0
openpyxl>=3.1.0
thop>=0.1.1
torchinfo>=1.8.0
medpy>=0.4.0
wandb>=0.16.0
gdown>=4.7.0
kaggle>=1.6.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
